Initial Commit + XGBoost example
First Commit of the HPOlib3
It contains the xgboost example.
This works locally or with a container. See examples.
Make sure to install the right dependencies. pip install .[xgboost] for
the local example or pip install .[xgboost,singularity] for the
container example. (maybe it is important not to have a space between
xgboost and singularity)