[skip travis] Fix#66 return by seed (#68)
* Return Sum of Runtimes. Not the mean.

* Update Nasbench101: Runindex
We can now specify a run index (0-2). The runindex is chosen in the original nasbench101 benchmark randomly.

* Reduce container size of nasbench101 container
Remove the api file from the container, instead download it before usage.
fix small bug in downloading procedure.

* Tabular Benchmark: Run Index
The runindex is set explicitly to None, then a random run id (one of 0,1,2,3) is selected.
That doesn't change the default behavior. When no run_index is specified, the run_index still is (0,1,2,3)

*  Update doc strings, Tests, Readme,

* Remove unnecessary dependencies

* Fix CodeStyle Error