Test: Availability test does not work for gitlab
The check if all container are available does not work out of the box for the gitlab registry. The gitlab registry does not support the `singularity search` function. At least not with the currently used version. But the container are also still available on the sylabs registry (as a backup). Fixing this tasks is still a TODO, but with lower priority.